Separate FakeGitHub into its own file
